src/main/java/com/gmail/at/sergey/christensen/helloWorldWithSpringSecurity/controllers/MainController.java
package com.gmail.at.sergey.christensen.helloWorldWithSpringSecurity.controllers;

import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.servlet.ModelAndView;

@Controller
public class MainController {

	@GetMapping(value = "/")
	public ModelAndView index() {
		ModelAndView modelAndView = new ModelAndView("index");
		return modelAndView;
	}
	
	@GetMapping(value = "/login")
	public ModelAndView login() {
		ModelAndView modelAndView = new ModelAndView("login");
		return modelAndView;
	}
	
	@GetMapping(value = "/registration")
	public ModelAndView registration() {
		ModelAndView modelAndView = new ModelAndView("registration");
		return modelAndView;
	}
	
	@GetMapping(value = "/accessDenied")
	public String accessDenied() {
		return "accessDenied";
	}
}
